+++ ui/lesson/InputChallenge.tsx
