+++ ui/lesson/TerminalChallenge.tsx
