+++ ui/lesson/index.ts
@@ -1,12 +1,12 @@
 import Lesson, { useLessonContext } from './Lesson'
 import LessonInfo from './Info'
-import LessonPrompt from './InputChallenge/Prompt'
-import LessonTerminal from './TerminalChallenge/Terminal'
+import LessonPrompt from './Prompt'
+import LessonTerminal from './Terminal'
 import LessonTabs from './Tabs'
-import TerminalChallenge from './TerminalChallenge/TerminalChallenge'
-import InputChallenge from './InputChallenge/InputChallenge'
+import TerminalChallenge from './TerminalChallenge'
+import InputChallenge from './InputChallenge'
 import ScriptingChallenge from './ScriptingChallenge'
-import HashChallenge from './HashChallenge/HashChallenge'
+import HashChallenge from './HashChallenge'
 
 export {
   Lesson,
